FIX: Fix installation order for hba file
In order to add data into hba file you have to install postgresql first.

+++ spec/spec_helper.rb
@@ -1,5 +1,5 @@
 # Based on https://github.com/puppetlabs/puppetlabs-ntp/blob/master/spec/spec_helper.rb
-# Thanks to Ken Barber for advice about http://projects.puppetlabs.com/issues/11191  
+# Thanks to Ken Barber for advice about http://projects.puppetlabs.com/issues/11191
 require 'puppet'
 require 'rspec-puppet'
 require 'tmpdir'
@@ -18,7 +18,9 @@ RSpec.configure do |c|
   c.filter_run_excluding :broken => true
 
   c.after :each do
-    FileUtils.remove_entry_secure(@puppetdir)
+    if FileTest.exists?("#{@puppetdir}")
+      FileUtils.remove_entry_secure(@puppetdir)
+    end
   end
 
   c.module_path = File.join(File.dirname(__FILE__), '../../')
